+++ src/printer.js
@@ -289,29 +289,22 @@ function handleJavaVariableExpression(path, print) {
   return concat(parts);
 }
 
-function handleLiteralExpression(path, print) {
+function handleLiteralExpression(path, print, options) {
+  const node = path.getValue();
   const literalType = path.call(print, "type", "$");
   if (literalType === "NULL") {
     return "null";
   }
-  let literalDoc = path.call(print, "literal", "$");
+  const literalDoc = path.call(print, "literal", "$");
   let doc;
   if (literalType === "STRING") {
     doc = concat(["'", literalDoc, "'"]);
-  } else if (literalType === "LONG") {
-    doc = concat([literalDoc, "L"]);
-  } else if (literalType === "DECIMAL" && literalDoc.indexOf(".") === -1) {
-    // The serializer does not preserve trailing 0s for doubles,
-    // however they are needed when printing,
-    literalDoc += ".0";
-  } else if (literalType === "DOUBLE") {
-    if (literalDoc.indexOf(".") === -1) {
-      // The serializer does not preserve trailing 0s for doubles,
-      // however they are needed when printing,
-      // e.g. 1.0D is valid while 1D is invalid
-      literalDoc += ".0";
-    }
-    doc = concat([literalDoc, "d"]);
+  } else if (
+    literalType === "LONG" ||
+    literalType === "DECIMAL" ||
+    literalType === "DOUBLE"
+  ) {
+    doc = options.originalText.slice(node.loc.startIndex, node.loc.endIndex);
   }
   if (doc) {
     return doc;