+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Fix incorrect format for decimals/doubles ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/88)).
 - Add support for Number Expression in SOQL Geolocation Literal ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/90)).
+- Fix duplicated trailing empty lines for multiple expressions on the same line ([issue](https://github.com/dangmai/prettier-plugin-apex/issues/91)).
 
 # 1.0.0-rc.5